Issue: https://gitee.com/openharmony/build/issues/I8GAP9 Comment: Fix Bug

+++ test/test_example_template/test_ohos_shared_library/include/hello.h
@@ -15,6 +15,6 @@
 #ifndef INCLUDE_HELLO_H
 #define INCLUDE_HELLO_H
 
-long fact(int);
+uint32_t fact(int);
 
 #endif  //  INCLUDE_HELLO_H
